// src/lib/enums/settings.enum.ts
export enum SettingsEnum{
    INSTRUCTIONS = "instructions",
    EXMWALL_LAST_UPDATED_AT = "exmwall_last_updated_at",
    CLIENT_URL= "client_url",
    SECTIONS = "sections",
    QUESTIONDOWNLOADTIME = "questiondownloadtime",
    QPKEYSDOWNLOADTIME = "qpkeysdownloadtime",
    EXAMDETAILSDOWNLOADTIME = "examdetailsdownloadtime"
}